+++ src/GoingMy.Web/src/app/models/post.model.ts
@@ -2,7 +2,6 @@ import { Author } from './author.model';
 
 export interface Post {
   id: string;
-  title: string;
   content: string;
   userId: string;
   username: string;
@@ -11,6 +10,7 @@ export interface Post {
   likes?: number;
   comments?: number;
   author?: Author;
+  userHasLiked?: boolean;
 }
 
 export interface Comment {